chore: split cns/service/main.go into logical files
No code changes, only rearranging. Break up the 1800-line main.go by
moving functions to files grouped by functional area:

- version.go: version string and print
- args.go: CLI argument definitions
- managed.go: managed DNC mode (registerNode, sendRegisterNodeRequest)
- aks.go: AKS/k8s initialization (InitializeCRDState, reconcileInitialCNSState, etc)
