+++ cns/service/args.go
@@ -1,0 +1,221 @@
+// Copyright 2017 Microsoft. All rights reserved.
+// MIT License
+
+package main
+
+import (
+	"os"
+
+	acn "github.com/Azure/azure-container-networking/common"
+	"github.com/Azure/azure-container-networking/log"
+	"github.com/Azure/azure-container-networking/platform"
+)
+
+// Command line arguments for CNS.
+var args = acn.ArgumentList{
+	{
+		Name:         acn.OptEnvironment,
+		Shorthand:    acn.OptEnvironmentAlias,
+		Description:  "Set the operating environment",
+		Type:         "string",
+		DefaultValue: acn.OptEnvironmentAzure,
+		ValueMap: map[string]interface{}{
+			acn.OptEnvironmentAzure:    0,
+			acn.OptEnvironmentMAS:      0,
+			acn.OptEnvironmentFileIpam: 0,
+		},
+	},
+	{
+		Name:         acn.OptAPIServerURL,
+		Shorthand:    acn.OptAPIServerURLAlias,
+		Description:  "Set the API server URL",
+		Type:         "string",
+		DefaultValue: "",
+	},
+	{
+		Name:         acn.OptLogLevel,
+		Shorthand:    acn.OptLogLevelAlias,
+		Description:  "Set the logging level",
+		Type:         "int",
+		DefaultValue: acn.OptLogLevelInfo,
+		ValueMap: map[string]interface{}{
+			acn.OptLogLevelInfo:  log.LevelInfo,
+			acn.OptLogLevelDebug: log.LevelDebug,
+		},
+	},
+	{
+		Name:         acn.OptLogTarget,
+		Shorthand:    acn.OptLogTargetAlias,
+		Description:  "Set the logging target",
+		Type:         "int",
+		DefaultValue: acn.OptLogTargetFile,
+		ValueMap: map[string]interface{}{
+			acn.OptLogTargetSyslog: log.TargetSyslog,
+			acn.OptLogTargetStderr: log.TargetStderr,
+			acn.OptLogTargetFile:   log.TargetLogfile,
+			acn.OptLogStdout:       log.TargetStdout,
+			acn.OptLogMultiWrite:   log.TargetStdOutAndLogFile,
+		},
+	},
+	{
+		Name:         acn.OptLogLocation,
+		Shorthand:    acn.OptLogLocationAlias,
+		Description:  "Set the directory location where logs will be saved",
+		Type:         "string",
+		DefaultValue: "",
+	},
+	{
+		Name:         acn.OptIpamQueryUrl,
+		Shorthand:    acn.OptIpamQueryUrlAlias,
+		Description:  "Set the IPAM query URL",
+		Type:         "string",
+		DefaultValue: "",
+	},
+	{
+		Name:         acn.OptIpamQueryInterval,
+		Shorthand:    acn.OptIpamQueryIntervalAlias,
+		Description:  "Set the IPAM plugin query interval",
+		Type:         "int",
+		DefaultValue: "",
+	},
+	{
+		Name:         acn.OptCnsURL,
+		Shorthand:    acn.OptCnsURLAlias,
+		Description:  "Set the URL for CNS to listen on",
+		Type:         "string",
+		DefaultValue: "",
+	},
+	{
+		Name:         acn.OptCnsPort,
+		Shorthand:    acn.OptCnsPortAlias,
+		Description:  "Set the URL port for CNS to listen on",
+		Type:         "string",
+		DefaultValue: "",
+	},
+	{
+		Name:         acn.OptVersion,
+		Shorthand:    acn.OptVersionAlias,
+		Description:  "Print version information",
+		Type:         "bool",
+		DefaultValue: false,
+	},
+	{
+		Name:         acn.OptStoreFileLocation,
+		Shorthand:    acn.OptStoreFileLocationAlias,
+		Description:  "Set store file absolute path",
+		Type:         "string",
+		DefaultValue: platform.CNMRuntimePath,
+	},
+	{
+		Name:         acn.OptNetPluginPath,
+		Shorthand:    acn.OptNetPluginPathAlias,
+		Description:  "Set network plugin binary absolute path to parent (of azure-vnet and azure-vnet-ipam)",
+		Type:         "string",
+		DefaultValue: platform.K8SCNIRuntimePath,
+	},
+	{
+		Name:         acn.OptNetPluginConfigFile,
+		Shorthand:    acn.OptNetPluginConfigFileAlias,
+		Description:  "Set network plugin configuration file absolute path",
+		Type:         "string",
+		DefaultValue: platform.K8SNetConfigPath + string(os.PathSeparator) + defaultCNINetworkConfigFileName,
+	},
+	{
+		Name:         acn.OptCreateDefaultExtNetworkType,
+		Shorthand:    acn.OptCreateDefaultExtNetworkTypeAlias,
+		Description:  "Create default external network for windows platform with the specified type (l2bridge)",
+		Type:         "string",
+		DefaultValue: "",
+	},
+	{
+		Name:         acn.OptTelemetry,
+		Shorthand:    acn.OptTelemetryAlias,
+		Description:  "Set to false to disable telemetry. This is deprecated in favor of cns_config.json",
+		Type:         "bool",
+		DefaultValue: true,
+	},
+	{
+		Name:         acn.OptHttpConnectionTimeout,
+		Shorthand:    acn.OptHttpConnectionTimeoutAlias,
+		Description:  "Set HTTP connection timeout in seconds to be used by http client in CNS",
+		Type:         "int",
+		DefaultValue: "5",
+	},
+	{
+		Name:         acn.OptHttpResponseHeaderTimeout,
+		Shorthand:    acn.OptHttpResponseHeaderTimeoutAlias,
+		Description:  "Set HTTP response header timeout in seconds to be used by http client in CNS",
+		Type:         "int",
+		DefaultValue: "120",
+	},
+	{
+		Name:         acn.OptPrivateEndpoint,
+		Shorthand:    acn.OptPrivateEndpointAlias,
+		Description:  "Set private endpoint",
+		Type:         "string",
+		DefaultValue: "",
+	},
+	{
+		Name:         acn.OptInfrastructureNetworkID,
+		Shorthand:    acn.OptInfrastructureNetworkIDAlias,
+		Description:  "Set infrastructure network ID",
+		Type:         "string",
+		DefaultValue: "",
+	},
+	{
+		Name:         acn.OptNodeID,
+		Shorthand:    acn.OptNodeIDAlias,
+		Description:  "Set node name/ID",
+		Type:         "string",
+		DefaultValue: "",
+	},
+	{
+		Name:         acn.OptManaged,
+		Shorthand:    acn.OptManagedAlias,
+		Description:  "Set to true to enable managed mode. This is deprecated in favor of cns_config.json",
+		Type:         "bool",
+		DefaultValue: false,
+	},
+	{
+		Name:         acn.OptDebugCmd,
+		Shorthand:    acn.OptDebugCmdAlias,
+		Description:  "Debug flag to retrieve IPconfigs, available values: assigned, available, all",
+		Type:         "string",
+		DefaultValue: "",
+	},
+	{
+		Name:         acn.OptDebugArg,
+		Shorthand:    acn.OptDebugArgAlias,
+		Description:  "Argument flag to be paired with the 'debugcmd' flag.",
+		Type:         "string",
+		DefaultValue: "",
+	},
+	{
+		Name:         acn.OptCNSConfigPath,
+		Shorthand:    acn.OptCNSConfigPathAlias,
+		Description:  "Path to cns config file",
+		Type:         "string",
+		DefaultValue: "",
+	},
+	{
+		Name:         acn.OptTelemetryService,
+		Shorthand:    acn.OptTelemetryServiceAlias,
+		Description:  "Flag to start telemetry service to receive telemetry events from CNI. Default, disabled.",
+		Type:         "bool",
+		DefaultValue: false,
+	},
+	{
+		Name:         acn.OptCNIConflistFilepath,
+		Shorthand:    acn.OptCNIConflistFilepathAlias,
+		Description:  "Filepath to write CNI conflist when CNI conflist generation is enabled",
+		Type:         "string",
+		DefaultValue: "",
+	},
+	{
+		Name:         acn.OptCNIConflistScenario,
+		Shorthand:    acn.OptCNIConflistScenarioAlias,
+		Description:  "Scenario to generate CNI conflist for",
+		Type:         "string",
+		DefaultValue: "",
+	},
+}
